refactored AssertDriverContents out of GraphView tests
